requirements.txt
absl-py==0.9.0
addict==2.2.0
aiohttp==3.6.2
aiohttp-cors==0.7.0
aioredis==1.3.1
apex==0.1
astor==0.8.1
async-timeout==3.0.1
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.3
bleach==3.1.5
blessings==1.7
cachetools==4.0.0
certifi==2021.5.30
cffi==1.12.3
chardet==3.0.4
click==7.1.2
cloudpickle==1.3.0
colorama==0.4.3
colorful==0.5.4
contextlib2==0.6.0.post1
contextvars==2.4
cycler==0.10.0
Cython==0.29.13
dataclasses @ file:///tmp/build/80754af9/dataclasses_1614363715916/work
decorator==4.4.2
defusedxml==0.6.0
dominate==2.3.5
easydict==1.9
editdistance==0.5.3
entrypoints==0.3
enum34==1.1.10
filelock==3.0.12
future==0.18.2
gast==0.2.2
gitdb==4.0.5
GitPython==3.1.9
google==3.0.0
google-api-core==1.22.4
google-auth==1.22.1
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.52.0
gpustat==0.6.0
grpcio==1.32.0
gym==0.17.1
h5py==2.9.0
hiredis==1.1.0
hyperopt @ git+https://github.com/hyperopt/hyperopt.git@6c5b33be93a71a4ae6a38f04a29b18020b78b702
idna==2.8
idna-ssl==1.1.0
imageio==2.8.0
imagesize==1.2.0
immutables==0.14
importlib-metadata==1.6.0
ipykernel==5.2.1
ipython==7.13.0
ipython-genutils==0.2.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.17.0
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
kornia==0.1.3.post2
lmdb==0.94
lxml==4.5.0
magickwand==0.2
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
msgpack==1.0.0
multidict==4.7.6
munch==2.5.0
natsort==7.1.1
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nltk==3.6.2
nomen==0.42
notebook==6.0.3
numpy==1.19.5
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
olefile==0.46
opencensus==0.7.10
opencensus-context==0.1.1
opencv-contrib-python==3.4.2.17
opencv-python==3.4.2.17
opt-einsum==3.2.0
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow @ file:///tmp/build/80754af9/pillow_1625649052827/work
pretrainedmodels==0.7.4
prometheus-client==0.7.1
prompt-toolkit==3.0.5
protobuf==3.12.0
psutil==5.7.2
ptyprocess==0.6.0
py-spy==0.3.3
py4j==0.10.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.3
pycparser==2.19
pyglet==1.5.0
Pygments==2.6.1
PyOpenGL==3.1.1a1
pyparsing==2.3.1
pyrsistent==0.16.0
pystopwatch2 @ git+https://github.com/ildoonet/pystopwatch2.git@64ecd49a2a164b665f44c54d7015b0f7d8497b7e
python-dateutil==2.8.0
python-speech-features==0.6
pytz==2019.3
PyWavelets==1.1.1
PyYAML==3.13
pyzmq==18.0.1
ray==1.0.0
redis==3.4.1
regex==2021.8.28
requests==2.21.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.15.0
scikit-learn==0.20.0
scipy==1.4.1
Send2Trash==1.5.0
six @ file:///tmp/build/80754af9/six_1623709665295/work
sklearn==0.0
smmap==3.0.4
soupsieve==2.0.1
straug==0.1.2
tensorboard==1.15.0
tensorboardX==1.4
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
theconf @ git+https://github.com/wbaek/theconf@de32022f8c0651a043dc812d17194cdfd62066e8
thop==0.0.31.post2005241907
threadpoolctl==2.1.0
tifffile==2020.5.25
torch==1.8.0
torchaudio==0.8.0a0+a751e1d
torchfile==0.1.0
torchvision==0.9.0
tornado==6.0.1
tqdm==4.50.2
traitlets==4.3.3
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3==1.24.1
visdom==0.1.8.8
Wand==0.6.7
warmup-scheduler @ git+https://github.com/ildoonet/pytorch-gradual-warmup-lr.git@08f7d5e5ac2f68f8c043b67d86b016df1bec34d8
warpctc-pytorch @ file:///home/liulab/lipu/softwares/warpctc_pytorch-0.2.1%2Btorch14.cuda102-cp36-cp36m-manylinux1_x86_64.whl
Wave==0.0.2
wavio==0.0.4
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.55.0
Werkzeug==1.0.0
wrapt==1.12.1
xml-python==0.3.5
yarl==1.6.0
zipp==3.1.0
